dockerpull.sh

+++ sh/dockerpull.sh
@@ -12,11 +12,13 @@ for IMG in $LST ; do
     IMGID=`docker images | grep "$DKRREP/$IDSTR/$IMG" | grep latest | awk -F" " '{print $3}'`
     docker pull $DKRREP/$IDSTR/$IMG 
     #
-    NONE=`docker images | grep $IMGID | grep "\<none\>"`
-    if [ "$NONE" != "" ]; then
-        TMPN="$DKRREP/$IDSTR/$IMG:del"
-        docker tag $IMGID $TMPN
-        docker rmi $TMPN
+    if [ "$IMGID" != "" ]; then
+        NONE=`docker images | grep $IMGID | grep "\<none\>"`
+        if [ "$NONE" != "" ]; then
+            TMPN="$DKRREP/$IDSTR/$IMG:del"
+            docker tag $IMGID $TMPN
+            docker rmi $TMPN
+        fi
     fi   
 done
